move to bqx, fix bugs, improve comments
